fix: fixed error display function

+++ include/ui.hpp
@@ -45,7 +45,7 @@ void printFoundDrives(std::vector<uint16_t> ids);
 void printUnableToFindCfgFile(std::string path);
 void printDriveInfoExtended(mab::Md80& drive, bool printAll);
 void printAllErrors(mab::Md80& drive);
-void printErrorDetails(uint16_t error, const std::vector<std::string>& errorList);
+void printErrorDetails(uint32_t error, const std::vector<std::string>& errorList);
 void printParameterOutOfBounds(std::string category, std::string field);
 void printFailedToSetupMotor(mab::Md80Reg_E regId);
